# PYTHON-4350 Faster and more consistent performance benchmark execution times