Reduce max benchmark time to 60 seconds and increase min time to 30 seconds.

Also introduces a new FASTBENCH setting to the benchmarks more quickly:
```
$ FASTBENCH=1 python test/performance/perf_test.py -v TestDeepEncoding
runTest (__main__.TestDeepEncoding.runTest) ... Completed TestDeepEncoding 179.236 MB/s, MEDIAN=0.011s, total time=0.100s, iterations=9
```
VS without:
```
$ python test/performance/perf_test.py -v TestDeepEncoding
runTest (__main__.TestDeepEncoding.runTest) ... Completed TestDeepEncoding 181.506 MB/s, MEDIAN=0.108s, total time=30.103s, iterations=277
```

https://jira.mongodb.org/browse/PYTHON-4350